Fixed zeroname_updater.py crashing underwindows with UTF8
For some reason windows doesn't like parsing namecoin domains with "weird" characters, so adding a try to catch exception stops a strange block crashing the script - instead it ignores and moves on.
We just have to hope there doesn't happen to be a namecoin domain in a block with weird characters.