main: logging: do not override the warn level
This fixes the very annoying problem where the log messages with
level WARNING are not printed, while log messages with lower
level INFO are printed. This is very confusing behavior and should
be avoided. This was due to this override. If there is too much
logging output in debug mode, then let's deescalate the level
of the frequently-printed log messages down to DEBUG level.