Added possibility to specify default and noclone files in the content.json
This commit ended up defining a different syntax than the hypothized one.
In particular the "default_files" key is an object and not a list.
The keys are the file sources and the values are files' destinations.

I give an usage example:
```
"default_files": {
  "config.ini.example": "config.ini",
  ...
},
"noclone_files": [
  "uninmportant-file.txt",
  ...
]
```

Also note that this only works for files.