imports: refer to modules via import objs
Replace the access via sys.modules (which is prone to violate
encapsulation).

For main, import strictly on demand (not at the top of each file),
because importing main has side-effects (unfortunately), and importing
it at time of loading the respective module breaks things (at the very
least, it breaks tests, due to argparser arguments getting created
twice -- once by test script and once by main).

Unfortunately, not all usage of sys.modules can be replaced with imports
at this point in time. Namely, TorManager.py and Site.py are not only
accessing main via sys.modules but also check whether main is loaded or
not. To fix this, a significant change to main is required -- to factor
it such that the owner of the singletons can be imported without
sideeffects.