Requests with `HOST` in `ui_host` shouldn't be proxy requests
Before this commit, all requests with non-IP `HOST` header were considered as proxy requests. This means that using ZeroNet as a public proxy with own domain would also be considered as proxy request, so all requests would fail.

This now checks if `HOST` is in `ui_host` config and if it is, it is considered as normal request and not as a proxy. This should now allow using ZeroNet on a public domain.