Update LICENSE
You are a piece of shit @imachug nobody voted for your shitty GLPv3! Those who voted they are not even surpassing the 13 votes! You don't pretend you given choice to people because that is bullshit! 

@HostFat asked: 

> Well, I'm ready I guess, what I need to do? :)
https://github.com/HelloZeroNet/ZeroNet/issues/2273#issuecomment-600106455 

@imachug replied: 
> Just post a comment "**GPLv3**" under this issue.
https://github.com/HelloZeroNet/ZeroNet/issues/2273#issuecomment-600107147

You and @shortcutme cheated (**illegally** manipulated) votes on a license change! Counted all "I don't care" as it is a vote for your preferred version of the license and when someone asked what to vote for you said again your preferred version and not that what is the most free or it is in the interest of the ZeroNet project! 

On the license change the all the "I don't care" votes MUST go to that which the community and not one person wanted the most! You can go to https://github.com/HelloZeroNet/ZeroNet/issues/2273 and count all the votes you will see clearly that everyone wanted **GLPv3+** (**GPLv3 or any later version**).

https://github.com/HelloZeroNet/ZeroNet/issues/2606#issuecomment-675306519
> It's true that I wanted to switch to a particular license

You are a piece of shit criminal just like Tamas Kocsis who included a backdoor in ZeroNet! When I come and demanded to remove the backdoor you accused me that I trying to add it! 

Link: https://github.com/HelloZeroNet/ZeroNet/issues/2606#issuecomment-675315211